+++ splunk/com/splunk/Args.java
@@ -18,7 +18,6 @@ package com.splunk;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
-import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -28,8 +27,8 @@ import java.util.Map.Entry;
  * arguments.
  *
  * This extension is used mainly for encoding arguments for UTF8 transmission
- * to a Splunk instance in a key?value pairing for a string, or
- * key?value1&key?value2 (and so on) for an array of strings.
+ * to a Splunk instance in a key=value pairing for a string, or
+ * key=value1&key=value2 (and so on) for an array of strings.
  */
 public class Args extends LinkedHashMap<String, Object> {